In order to focus on definitions content, removed 'typings' from the toolchain. See deployment related issue #1. As a result, imports are now set to relative paths in the interim to postpone decision regarding deployment and module resolution.
Updates to d3-selection and d3-transition to replace use of type aliases with direct argument typing of function arguments passed into methods or returned. This is done with regard to ease visibility of typing info in editor without need to navigate to alias definition (development is in VS Code). Some updates to naming of generics (e.g. DescElement to generalize beyon immediate  child elements in selections, also use OldDatum in selectAll, where pre-existing data may be selected with the elements.)